+++ src/tables/cmap/format0.rs
@@ -20,9 +20,30 @@ pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
     }
 }
 
+pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) -> Option<()> {
+    let mut s = Stream::new(data);
+    s.skip::<u16>(); // format
+    s.skip::<u16>(); // length
+    s.skip::<u16>(); // language
+
+    for code_point in 0..256 {
+        // In contrast to every other format, here we take a look at the glyph
+        // id and check whether it is zero because otherwise this method would
+        // always simply call `f` for `0..256` which would be kind of pointless
+        // (this array always has length 256 even when the face has only fewer
+        // glyphs).
+        let glyph_id: u8 = s.read()?;
+        if glyph_id != 0 {
+            f(code_point);
+        }
+    }
+
+    Some(())
+}
+
 #[cfg(test)]
-mod format0_tests {
-    use super::parse;
+mod tests {
+    use super::{parse, codepoints};
 
     #[test]
     fn maps_not_all_256_codepoints() {
@@ -39,5 +60,9 @@ mod format0_tests {
         assert_eq!(parse(&data, 0), None);
         assert_eq!(parse(&data, 0x40), Some(100));
         assert_eq!(parse(&data, 100), None);
+
+        let mut vec = vec![];
+        codepoints(&data, |c| vec.push(c));
+        assert_eq!(vec, [0x40]);
     }
 }