+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `cmap::Subtable::codepoints`
+
 ### Fixed
 - (cmap) Incorrectly returning glyph ID `0` instead of `None` for format 0
 - (cmap) Possible invalid glyph mapping for format 2